requirements.txt
#faiss-cpu==1.11.0
torch==2.7.1
pandas==2.3.1
transformers==4.53.1
hdbscan==0.8.40
fastparquet==2024.11.0
sentence-transformers==5.1.0
protobuf==6.32.0
faiss-gpu==1.7.2
numpy==1.26.4
diskcache==5.6.3
